chainged map and switched to new index
