+++ src/county-selecter.js
@@ -3,7 +3,7 @@ import React from 'react'
 const CountySelector=({props})=>{
 
     return <div>
-        
+        {props}
     </div>
 
 }
