+++ src/nav.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Nav=({props})=>{
+     return <h1>{props.value}</h1>; 
+}
+
+
+
+export default Nav;
